Globalfucntion use the reference module to get the datacenter and filecenter info. Manager add template type response.

+++ src/modules/core/result.h
@@ -126,7 +126,7 @@ namespace Sloong
         T m_tResultObject;
     };
 
-    typedef TResult<int> NResult;
+    typedef TResult<int64_t> NResult;
 	typedef TResult<std::pair<char*,int>> PResult;
     typedef TResult<string> SResult;
     typedef TResult<vector<string>> VStrResult;
